+++ src/domain/mediaTypes.js
@@ -1,0 +1,16 @@
+/**
+ * @module domain/mediaTypes
+ */
+/**
+ * Enum for common media types.
+ * @readonly
+ * @enum {String}
+ */
+export const MediaType = Object.freeze({
+	JSON: 'application/json',
+	TEXT_HTML: 'text/html',
+	TEXT_PLAIN: 'text/plain',
+	KML: 'application/vnd.google-earth.kml+xml',
+	GPX: 'application/gpx+xml',
+	GeoJSON: 'application/geo+json'
+});
