+++ src/modules/geoResourceInfo/services/provider/geoResourceInfoResult.provider.js
@@ -3,7 +3,8 @@
  */
 import { $injector } from '../../../../injection';
 import { GeoResourceInfoResult } from '../GeoResourceInfoService';
-import { MediaType } from '../../../../services/HttpService';
+import { MediaType } from '../../../../domain/mediaTypes';
+
 import { GeoResourceAuthenticationType, WmsGeoResource } from '../../../../domain/geoResources';
 
 /**
