+++ src/services/HttpService.js
@@ -129,17 +129,3 @@ export class NetworkStateSyncHttpService extends HttpService {
 		}
 	}
 }
-
-/**
- * Enum for common media types.
- * @readonly
- * @enum {String}
- */
-export const MediaType = Object.freeze({
-	JSON: 'application/json',
-	TEXT_HTML: 'text/html',
-	TEXT_PLAIN: 'text/plain',
-	KML: 'application/vnd.google-earth.kml+xml',
-	GPX: 'application/gpx+xml',
-	GeoJSON: 'application/geo+json'
-});
