+++ src/services/provider/mfp.provider.js
@@ -2,7 +2,8 @@
  * @module services/provider/mfp_provider
  */
 import { $injector } from '../../injection';
-import { HttpService, MediaType } from '../HttpService';
+import { HttpService } from '../HttpService';
+import { MediaType } from '../../domain/mediaTypes';
 
 /**
  * @typedef {Object} BvvMfpCapabilities
