+++ test/modules/export/components/dialog/ExportDialogContent.test.js
@@ -2,7 +2,7 @@ import { SourceTypeName } from '../../../../../src/domain/sourceType';
 import { $injector } from '../../../../../src/injection';
 import { ExportDialogContent } from '../../../../../src/modules/export/components/dialog/ExportDialogContent';
 import { ExportItem } from '../../../../../src/modules/export/components/dialog/ExportItem';
-import { MediaType } from '../../../../../src/services/HttpService';
+import { MediaType } from '../../../../../src/domain/mediaTypes';
 import { TestUtils } from '../../../../test-utils';
 
 window.customElements.define(ExportDialogContent.tag, ExportDialogContent);
