+++ test/service/HttpService.test.js
@@ -1,4 +1,4 @@
-import { MediaType, HttpService, NetworkStateSyncHttpService } from '../../src/services/HttpService';
+import { HttpService, NetworkStateSyncHttpService } from '../../src/services/HttpService';
 import { networkReducer } from '../../src/store/network/network.reducer';
 import { TestUtils } from '../test-utils';
 
@@ -308,18 +308,5 @@ describe('NetworkStateSyncHttpService', () => {
 				expect(store.getState().network.fetching).toBeFalse();
 			}
 		});
-	});
-});
-
-describe('MediaType', () => {
-	it('is an enum representing common media types', () => {
-		expect(Object.entries(MediaType).length).toBe(6);
-		expect(Object.isFrozen(MediaType)).toBeTrue();
-		expect(MediaType.JSON).toEqual('application/json');
-		expect(MediaType.TEXT_HTML).toEqual('text/html');
-		expect(MediaType.TEXT_PLAIN).toEqual('text/plain');
-		expect(MediaType.KML).toEqual('application/vnd.google-earth.kml+xml');
-		expect(MediaType.GPX).toEqual('application/gpx+xml');
-		expect(MediaType.GeoJSON).toEqual('application/geo+json');
 	});
 });
