+++ test/service/ImportVectorDataService.test.js
@@ -1,7 +1,7 @@
 import { $injector } from '../../src/injection';
 import { VectorGeoResource, VectorSourceType } from '../../src/domain/geoResources';
 import { SourceType, SourceTypeName, SourceTypeResult, SourceTypeResultStatus } from '../../src/domain/sourceType';
-import { MediaType } from '../../src/services/HttpService';
+import { MediaType } from '../../src/domain/mediaTypes';
 import { ImportVectorDataService } from '../../src/services/ImportVectorDataService';
 import { TestUtils } from '../test-utils';
 import {
