+++ test/service/mfp.provider.test.js
@@ -1,5 +1,6 @@
 import { $injector } from '../../src/injection';
-import { HttpService, MediaType } from '../../src/services/HttpService';
+import { HttpService } from '../../src/services/HttpService';
+import { MediaType } from '../../src/domain/mediaTypes';
 import { getMfpCapabilities, postMfpSpec } from '../../src/services/provider/mfp.provider';
 describe('mfp provider', () => {
 	describe('getMfpCapabilities', () => {
